excluded SlicedParagraph in a seperate component
