+++ apps/admin/src/components/Table/SlicedParagraph/SlicedParagraph.module.scss
@@ -1,0 +1,5 @@
+.showMore {
+  color: purple;
+  cursor: pointer;
+  text-decoration: underline;
+}
