+++ apps/admin/src/components/Table/SlicedParagraph/index.ts
@@ -1,0 +1,3 @@
+import SlicedParagraph from './SlicedParagraph';
+
+export default SlicedParagraph;
