+++ apps/admin/src/components/Table/Table.module.scss
@@ -17,12 +17,6 @@
   background-color: #fff;
 }
 
-.showMore {
-  color: purple;
-  cursor: pointer;
-  text-decoration: underline;
-}
-
 .actions {
   display: flex;
   gap: 4px;
